requirements.txt
dash>=2.14.0
dash-bootstrap-components>=1.5.0
matplotlib>=3.7.0
plotly>=5.17.0
Pillow>=10.0.0
pandas>=2.0.0
numpy>=1.24.0
pyanatomogram>=0.1.0
lifelines>=0.27.0
dash-ag-grid>=0.3.0
